✨ 강재수불부 집계 간산 로직 추가

+++ src/main/java/com/lineinc/erp/api/server/domain/steelmanagementv2/entity/SteelManagementDetailV2.java
@@ -106,14 +106,14 @@ public class SteelManagementDetailV2 extends BaseEntity {
      */
     @DiffInclude
     @Column
-    private Integer unitPrice;
+    private Long unitPrice;
 
     /**
      * 금액
      */
     @DiffInclude
     @Column
-    private Integer amount;
+    private Long amount;
 
     /**
      * 구분: 자사자재, 구매, 임대
